+++ .tarpaulin.toml
@@ -1,0 +1,9 @@
+[all]
+exclude-files = ["src/existing_types_in_sargon.rs"]
+verbose = false
+timeout = "2m"
+force-clean = true
+locked = true
+all-features = true
+jobs = 1
+out = ["Xml"]
