+++ src/types/mod.rs
@@ -1,7 +1,6 @@
 mod drivers;
 mod new_types;
 mod petitions;
-mod sargon;
 mod signatures_building_coordinator;
 mod signatures_outcome;
 mod signing_driver;
@@ -11,7 +10,6 @@ mod simulated_user;
 pub use drivers::*;
 pub use new_types::*;
 pub use petitions::*;
-pub use sargon::*;
 pub use signatures_building_coordinator::*;
 pub use signatures_outcome::*;
 pub use signing_driver::*;
