+++ src/types/new_types.rs
@@ -20,6 +20,55 @@ pub struct HDSignature {
     pub owned_factor_instance: OwnedFactorInstance,
 }
 
+#[derive(Clone, Debug, PartialEq, Eq, std::hash::Hash)]
+pub struct InvalidTransactionIfSkipped {
+    pub intent_hash: IntentHash,
+    pub entities_which_would_fail_auth: Vec<AccountAddressOrIdentityAddress>,
+}
+impl InvalidTransactionIfSkipped {
+    pub fn new(
+        intent_hash: IntentHash,
+        entities_which_would_fail_auth: Vec<AccountAddressOrIdentityAddress>,
+    ) -> Self {
+        Self {
+            intent_hash,
+            entities_which_would_fail_auth,
+        }
+    }
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, std::hash::Hash)]
+pub struct OwnedMatrixOfFactorInstances {
+    pub address_of_owner: AccountAddressOrIdentityAddress,
+    pub matrix: MatrixOfFactorInstances,
+}
+impl OwnedMatrixOfFactorInstances {
+    pub fn new(
+        address_of_owner: AccountAddressOrIdentityAddress,
+        matrix: MatrixOfFactorInstances,
+    ) -> Self {
+        Self {
+            address_of_owner,
+            matrix,
+        }
+    }
+}
+
+#[derive(Clone, PartialEq, Eq, std::hash::Hash, derive_more::Debug)]
+#[debug("{:?}: {:?}", owner, factor_instance)]
+pub struct OwnedFactorInstance {
+    pub factor_instance: FactorInstance,
+    pub owner: AccountAddressOrIdentityAddress,
+}
+impl OwnedFactorInstance {
+    pub fn new(factor_instance: FactorInstance, owner: AccountAddressOrIdentityAddress) -> Self {
+        Self {
+            factor_instance,
+            owner,
+        }
+    }
+}
+
 impl HDSignature {
     pub fn new(
         intent_hash: IntentHash,
